little fix scorpion movement

+++ core/character.c
@@ -8,6 +8,6 @@ void character_init(character_t* character, sprite_t* sprite, vector2_t speed)
 
 void character_tick(character_t* character, double delta_time)
 {
-    character->sprite->x += (character->speed.x * delta_time);
-    character->sprite->y += (character->speed.y * delta_time);
+    character->sprite->x += (int) (character->speed.x * delta_time);
+    character->sprite->y += (int) (character->speed.y * delta_time);
 }
